+++ tests/integration/harvest_job_flows/test_harvest_job_full_flow.py
@@ -12,7 +12,6 @@ from harvester.utils.general_utils import download_file
 HARVEST_SOURCE_URL = os.getenv("HARVEST_SOURCE_URL")
 
 
-@patch("harvester.lib.cf_handler.CloudFoundryClient")
 class TestHarvestJobFullFlow:
     @patch("harvester.harvest.ckan_sync_tool.ckan")
     @patch("harvester.harvest.HarvestSource.send_notification_emails")
@@ -386,7 +385,6 @@ class TestHarvestJobFullFlow:
 
     def test_harvest_multiple_jobs(
         self,
-        CFClientMock,  # Class-level patch parameter
         interface,
         organization_data,
         source_data_dcatus_single_record,
@@ -408,40 +406,6 @@ class TestHarvestJobFullFlow:
         harvest_job = interface.get_harvest_job(harvest_job1.id)
         assert f"Job {harvest_job2.id} is already in progress for source" in caplog.text
         assert harvest_job.status == "error"
-
-    def test_harvest_multiple_tasks(
-        self,
-        CFClientMock,  # Class-level patch parameter comes first
-        interface,
-        organization_data,
-        source_data_dcatus_single_record,
-        caplog,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus_single_record)
-        harvest_job = interface.add_harvest_job(
-            {
-                "status": "in_progress",
-                "harvest_source_id": source_data_dcatus_single_record["id"],
-            }
-        )
-
-        CFClientMock.return_value.v3.apps._pagination.return_value = [
-            {"state": "RUNNING", "name": f"harvest-job-{harvest_job.id}-harvest"},
-            {"state": "RUNNING", "name": f"harvest-job-{harvest_job.id}-harvest"},
-            {
-                "state": "RUNNING",
-                "name": "harvest-job-1c3d686c-6156-429d-b27b-5ab163750e76-harvest",
-            },
-        ]
-
-        caplog.set_level(logging.INFO)
-
-        harvest_job_starter(harvest_job.id, "harvest")
-        assert (
-            f"Job {harvest_job.id} is already running in another task. Exiting."
-            in caplog.text
-        )
 
 
 class TestCheckMoreWork:
